hc/requirements.txt
APScheduler==3.6.3
beautifulsoup4==4.11.2
Django==4.1.4
django-admin-list-filter-dropdown==1.0.3
django-constance==2.9.1
django-json-widget==1.1.1
django-picklefield==3.1
django-tgbot==2.0.0
openpyxl==3.0.10
psycopg2-binary==2.9.5
requests==2.28.1